src/gcf_extract_load_football_data_co_uk/requirements.txt
PyYAML==6.0
requests==2.27.1
google-api-core==2.7.1
google-api-python-client==2.42.0
google-cloud-storage==2.2.1
